refactoring/117 리팩토링

+++ src/main/java/com/select/choice/domain/comment/service/Impl/CommentServiceImpl.java
@@ -31,9 +31,7 @@ public class CommentServiceImpl implements CommentService {
         User user = userUtil.currentUser();
         Post post = postUtil.findById(postIdx);
         Comment comment = commentConverter.toEntity(commentDto, user, post);
-
         commentRepository.save(comment);
-        post.updateCount();
     }
 
     @Override
@@ -51,10 +49,12 @@ public class CommentServiceImpl implements CommentService {
     @Override
     @Transactional(rollbackFor = Exception.class)
     public void delete(Long postIdx, Long commentIdx) {
+        User user = userUtil.currentUser();
         Comment comment = commentUtil.findById(commentIdx);
-        Post post = postUtil.findById(postIdx);
 
-        post.minusCount();
+        if(!comment.getUser().equals(user)){
+            throw new IsNotMyCommentException(ErrorCode.IS_NOT_MY_COMMENT);
+        }
         commentRepository.delete(comment);
     }
 }
